+++ firmware/main/Control/Persistency/NvsStorageDriver.h
@@ -31,7 +31,7 @@ public:
     /**
      * @see ConfigProviderIfc::writeConfiguration
      */
-    virtual esp_err_t writeConfiguration(const char* name, const void* const* data, const size_t size) override;
+    virtual esp_err_t writeConfiguration(const char* name, const void* data, const size_t size) override;
 };
